requirements.txt
quart
hypercorn
python-dotenv
beautifulsoup4
jinja2
aiohttp
asyncio
requests